Split notebook utils into separate file
